factor out forking. fix property printing. usePolling with chokidar.
